### 2024-12-17 #7599 @editorialbot
[PRE REVIEW]: Balsa: A Fast C++ Random Forest Classifier with Commandline and Python Interface
**Submitting author:** @tobiasborsdorff (<a href="http://orcid.org/0000-0002-4421-0187">Tobias Borsdorff</a>)
**Repository:** https://github.com/SRON-Earth/Balsa
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
