### 2025-01-03 #7599 @crvernon
:wave: can you take this one on @HaoZeke?
