### 2025-01-16 #7599 @tobiasborsdorff
> @editorialbot invite @HaoZeke as editor
>
> 👋 can you take this one on @HaoZeke?

Dear @crvernon,

I hope this message finds you well. I’m writing to kindly follow up on the paper I submitted over a month ago. It seems that the editor assignment has not yet been initiated. Could you please let me know when this step might take place?

I truly appreciate your time and assistance and look forward to your response.

Best regards, Tobias Borsdorff
