### 2025-01-17 #7599 @crvernon
:wave: @tobiasborsdorff - we have a large backlog of submissions right now so it may take a little longer than usual to get you set up with an editor.  Thanks for your patience!

@HaoZeke - are you able to take this one on?
