### 2025-02-10 #7599 @dostuffthatmatters
Hi @HaoZeke,

Thank you for asking. I am busy with another review right now, so I have to kindly decline.

**A general remark:**

The main statement of need for this tool is about performance. But there is no performance evaluation against existing tools in the paper. The ScikitLearn implementations for decision trees and random forests are extremely efficient and there are many ML libraries like ScikitLearn with interfaces for different programming languages.

Maybe there is a good reason for the existence of Balsa, but I could not tell that from the given material. Good luck with the review!
