### 2025-02-10 #7599 @bcjaeger
> Hi [@bcjaeger](https://github.com/bcjaeger) 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

Hello! 👋 This looks very interesting, but I don't have enough availability to review at the moment. I agree that pointing to a formal benchmark of computational efficiency in the article would be a great addition.
