### 2025-04-07 #7599 @tobiasborsdorff
Dear HaoZeke,

Thank you for your feedback. We have already conducted a performance analysis of the Balsa algorithm, which is documented in Section 6.1 of the ATBD  https://zenodo.org/records/14186320 that is also referenced in our paper.

As illustrated by the figures in that chapter, Balsa demonstrates clear advantages in both memory usage and runtime compared to the scikit-learn implementation as well as ranger—a C++ implementation of the random forest algorithm.

We would be happy to include a brief summary of these results in the manuscript to enhance clarity. However, we kindly ask that this addition be addressed during the official review process. While @dostuffthatmatters raised a valuable point, it is important to note that the suggestion was made outside the formal review framework and the manuscript is still not under review.

Best regards, Tobias Borsdorff
