### 2025-04-09 #7599 @tobiasborsdorff
Dear @HaoZeke and @dostuffthatmatters,

I hope this message finds you well. I wanted to inform you that I have updated the manuscript to include a full paragraph including figures discussing the runtime and memory usage of the Balsa implementation in comparison with the SKlearn - Python and Ranger C++ implementations. This addition highlights the performance advantages of Balsa, particularly in terms of memory efficiency and prediction speed making the software suitability for large-scale operations.

With this update, I believe the manuscript is now complete and ready for the next steps. As the manuscript has been awaiting review since December last year, I would greatly appreciate it if you could kindly restart the review process and initiate the search for reviewers at your earliest convenience.

Thank you for your time and consideration.

Best regards, Tobias
