### 2025-04-14 #7599 @tobiasborsdorff
> Thanks [@tobiasborsdorff](https://github.com/tobiasborsdorff) , if you have any suggested reviewers please let me know without the @ here.

Dear @HaoZeke, I don’t know many people who review for JOSS, but a colleague of mine suggested these two: https://github.com/mkhorton and https://github.com/dgasmith. Maybe they’re a good fit! It’d be great if you could also reach out to a few people you think might be interested. regards Tobias
