### 2025-08-08 #7599 @vipinagrawal25
Hello @HaoZeke,

I can see that you are struggling in finding reviewer for this article. I may have a suggestion @bhaveshshrimali --  perhaps you could reach out to him. He has worked in the field of Computational Mechanics and Machine Learning, and has also open-sourced his work, most recently on Neural Operators. He also has also worked on, and deployed Tree-based models (Random Forests/Gradient-boosted Decision Trees) for his different works.
