### 2025-08-11 #8778 @editorialbot
[REVIEW]: Balsa: A Fast C++ Random Forest Classifier with Command-line and Python Interface
**Submitting author:** @tobiasborsdorff (<a href="http://orcid.org/0000-0002-4421-0187">Tobias Borsdorff</a>)
**Repository:** https://github.com/SRON-Earth/Balsa
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
