### 2025-09-06 #8778 @HaoZeke
@vipinagrawal25 for the unchecked items, could you open issues upstream so @tobiasborsdorff can address them?
