### 2025-09-16 #8778 @vipinagrawal25
Hello @tobiasborsdorff,

I would like to reproduce Figures 1 and 2 from the paper. Could you please confirm whether the corresponding tests for these figures have been added? I am unable to locate them.

--Vipin
