### 2025-09-26 #8778 @vipinagrawal25
Dear @HaoZeke ,

Thanks for offering me the opportunity to review the paper. First of all, congratulations on developing Balsa. It is a space and time efficient C++ random forest classifier. I believe that the software community could be benefitted from it. However I have following comments. Once they are addressed I recommend this paper for publication.

**Contribution and authorship**:- The lead author does not have the highest number of commits. 4th and 5th author Soumyajit Mandal, Jochen Landgraf, does not have any commits at all.

**Installation**:-  I can not run ``make install'' because install target is not defined in _CmakeLists.txt_. Can you please update the cmake so makefile can install balsa, balsa_train, balsa_generate etc.

**Performance**:- I have not been able to test the performance as the authors did not provide test cases to do so.

**Example usage**- The authors have provided Examples subdirectories. I am able to run them to generate, train and classify the data. The file does not work initially because balsa_train ... are not added into the path. But once the user fixes that everything works fine.

**Functionality**- Python Interface: Absent—no bindings, pip-installable package, or examples. Paper claims "versatile Python interface... installed with pip," but repo has none (discrepancy noted in borsdorff mirror too).

**Paper**:
- The title of second section. The line after Balsa should be in the description.
- Figure Figure are appearing twice everywhere.

**Other**:
Balsa appears to a common software which is already available in pip. Authors may think of another name if they wish to.

Addressing these points will strengthen the clarity and impact of Balsa, ensuring it is recognized as a valuable contribution to the open-source software community.
