### 2025-10-17 #8778 @tobiasborsdorff
Thank you very much, @bhaveshshrimali and @vipinagrawal25, for taking the time to review our submission and for your constructive and helpful feedback.
We will now proceed to update both the software and the paper to address your comments.
To ensure we correctly fulfill all requirements, we would like to ask @bhaveshshrimali to please specify what updates are needed for the unchecked boxes in the review checklist.
This would help us understand more precisely what needs to be improved — similar to the clarifications already provided by @vipinagrawal25 — since a few points are currently not entirely clear to us.
Once we have your clarification, we will implement the updates and summarise the changes here for your review. Best regards Tobias Borsdorff
