### 2025-11-10 c834473
Remove subfolder LICENSE files
