### 2025-11-10 3f75d81
Update README.md with main repository guide
