+++ tools/rfcperf/README.md
@@ -1,0 +1,317 @@
+![Balsa Logo](../../core/Images/BalsaLogo.png)
+
+# rfcperf: Random Forest Classifier Performance Tool
+
+A comprehensive benchmarking and profiling tool for comparing Random Forest classifier implementations across different dataset sizes and configurations.
+
+## Overview
+
+rfcperf is designed to systematically evaluate and compare Random Forest classifiers by:
+- Running training and classification across multiple dataset sizes
+- Measuring system performance (CPU time, memory usage, wall-clock time)
+- Calculating classification quality metrics (accuracy, precision, recall, etc.)
+- Generating comparative visualization reports
+
+This tool is particularly useful for:
+- Benchmarking different Random Forest implementations (Balsa, scikit-learn, etc.)
+- Optimizing classifier parameters for your dataset
+- Understanding scalability and resource usage patterns
+- Comparing trade-offs between speed, memory, and accuracy
+
+## Features
+
+- **Multi-format dataset support**: JSON, CSV, binary, and Balsa native formats
+- **Automated dataset generation**: Creates training/test splits at multiple sizes
+- **Comprehensive metrics**: System performance (time, memory) and classification quality (accuracy, F-scores, etc.)
+- **Visual reports**: Multi-page PDF reports with comparison graphs
+- **Caching**: Efficient dataset reuse across runs
+- **Configurable**: External configuration file for classifier settings
+- **Extensible**: Driver-based architecture for adding new classifiers
+
+## Installation
+```bash
+cd tools
+pip install -e .
+```
+
+**Requirements:**
+- Python ≥ 3.7
+- NumPy
+- matplotlib
+- jsonpickle
+
+## Quick Start
+
+### 1. Prepare Your Dataset
+
+Your training data should be in JSON format (using jsonpickle):
+```python
+import numpy as np
+import jsonpickle
+import jsonpickle.ext.numpy
+
+jsonpickle.ext.numpy.register_handlers()
+
+# Your data: 2D array of features (float32)
+data_points = np.array([[1.0, 2.0], [3.0, 4.0], ...], dtype=np.float32)
+
+# Your labels: 1D array (float32, values 0 or 1 for binary classification)
+labels = np.array([0, 1, 0, ...], dtype=np.float32)
+
+# Save as JSON
+with open("my_dataset.json", "w") as f:
+    f.write(jsonpickle.encode((data_points, labels)))
+```
+
+### 2. Generate Default Configuration
+```bash
+python -m rfcperf profile my_dataset.json balsa
+```
+
+This creates `rfcperf.ini` with default settings. Edit this file to configure your classifiers.
+
+### 3. Configure Classifiers
+
+Edit `rfcperf.ini`:
+```ini
+[rfcperf]
+cache_dir = cache
+run_dir = run
+
+[balsa]
+driver = balsa
+balsa_train = /path/to/balsa_train
+balsa_classify = /path/to/balsa_classify
+balsa_measure = /path/to/balsa_measure
+
+[sklearn]
+driver = sklearn
+# sklearn-specific options...
+```
+
+### 4. Run Profiling
+```bash
+python -m rfcperf profile my_dataset.json balsa sklearn \
+    -n 1000,5000,10000,50000 \
+    -e 150 \
+    -t 4 \
+    -p 33
+```
+
+This profiles both Balsa and scikit-learn on datasets of 1K, 5K, 10K, and 50K samples, using 150 trees, 4 threads, and a 33% test split.
+
+### 5. View Results
+
+Results are saved in timestamped directories under `run/`:
+```
+run/2025-01-15T10:30:00/
[…]
