### 2025-11-11 5990b1e
Add install targets to CMake and update README for custom installation paths
