### 2025-11-11 a634737
updating the pyproject.toml file to copy the license file
