### 2025-11-11 779bd36
CMakeLists.txt supports custom installation paths
