### 2025-11-12 da591e7
updated readme for the balsa python binding
