### 2025-11-12 8734036
adding an example for performance analysis
