+++ tools/rfcperf/README.md
@@ -28,39 +28,110 @@ This tool is particularly useful for:
 - **Configurable**: External configuration file for classifier settings
 - **Extensible**: Driver-based architecture for adding new classifiers
 
-## Installation
-```bash
-cd tools
-pip install -e .
-```
-
 **Requirements:**
 - Python ≥ 3.7
 - NumPy
 - matplotlib
 - jsonpickle
 
+
 ## Quick Start
 
 ### 1. Prepare Your Dataset
 
-Your training data should be in JSON format (using jsonpickle):
+#### Create Dataset Specification
+
+First, create a dataset specification file named `fruit.conf`:
+```
+multisource(4)
+{
+    source(78)
+    {
+        gaussian(122, 11  );
+        gaussian(40 , 9   );
+        gaussian(13 , 12  );
+        uniform( 100, 150 );
+    }
+    source(22)
+    {
+        gaussian(100, 10  );
+        gaussian(100, 10  );
+        gaussian( 20, 10  );
+        uniform( 110, 155 );
+    }
+}
+```
+
+This specification defines a binary classification problem with 4 features per data point. The two sources represent different classes (e.g., apples and oranges) with a 78:22 distribution ratio.
+
+#### Generate Dataset Using balsa_generate
+
+Use the `balsa_generate` tool to create structured random test data:
+```bash
+# Generate 30 million data points based on fruit.conf specification
+balsa_generate -p 30000000 fruit.conf fruit-points.balsa fruit-labels.balsa
+
+# Convert binary format to readable text files
+balsa_print fruit-points.balsa > fpoints.txt
+balsa_print fruit-labels.balsa > flabels.txt
+```
+
+#### Convert to JSON Format
+
+Once you have the text files, convert them to JSON format for use with the Python bindings:
 ```python
 import numpy as np
 import jsonpickle
-import jsonpickle.ext.numpy
 
-jsonpickle.ext.numpy.register_handlers()
+# Paths to the input text files
+points_file = "fpoints.txt"
+labels_file = "flabels.txt"
+output_json = "my_dataset.json"
 
-# Your data: 2D array of features (float32)
-data_points = np.array([[1.0, 2.0], [3.0, 4.0], ...], dtype=np.float32)
+def read_points(filename):
+    """Read floating point table from text file."""
+    data = []
+    with open(filename, "r") as f:
+        lines = f.readlines()
+    # Skip header lines until we find table rows
+    for line in lines:
+        line = line.strip()
+        if not line or line[0].isdigit() == False:
+            continue
+        # Row format: "0   : 123.303  43.3323  -5.20379 102.836"
+        parts = line.split(":")
+        values = parts[1].strip().split()
+        data.append([float(v) for v in values])
+    return np.array(data, dtype=np.float32)
 
-# Your labels: 1D array (float32, values 0 or 1 for binary classification)
-labels = np.array([0, 1, 0, ...], dtype=np.float32)
+def read_labels(filename):
+    """Read label table from text file."""
+    labels = []
+    with open(filename, "r") as f:
+        lines = f.readlines()
+    for line in lines:
+        line = line.strip()
+        if not line or line[0].isdigit() == False:
+            continue
+        # Row format: "0   : 0"
+        parts = line.split(":")
+        labels.append(int(parts[1].strip()))
+    return np.array(labels, dtype=np.float64)
 
-# Save as JSON
-with open("my_dataset.json", "w") as f:
-    f.write(jsonpickle.encode((data_points, labels)))
[…]
