### 2025-11-12 90313e4
adding an example for performance analysis parameter update
