+++ python/README.md
@@ -1,7 +1,9 @@
 ![](images/BalsaLogo.png)
+
 # Balsa: A Fast Random Forest Classifier
 
 ## About
+
 Balsa is a fast and memory-efficient implementation of the random forest
 classification algorithm. Balsa is optimized for low memory usage and high
 speed during both training and classification. It is particularly useful for
@@ -13,241 +15,80 @@ classification-optimization. This file covers the specifics of using the Python
 binding.
 
 ## Prerequisites
+
 The balsa-python package requires NumPy >= 1.25.
 
 ## Installation
+
 Balsa can be installed directly from a cloned Git repository, as follows:
-```bash
+
+```
 pip install .
 ```
+
 Installation of the balsa Python package using `pip` requires a working C/C++
 compiler, git, CMake (>=3.18), Python (>=3.9), NumPy (>= 1.25), and the Python
 development headers and static library (on Ubuntu the required package is
 called `python3-dev`).
 
-## Quick Start
+## Examples
 
-### Basic Training and Classification
-The following example demonstrates training and classification with default settings:
+### Training and Classification with Default Settings
 
-```python
+The following example code demonstrates how the Balsa trainer and classifier can
+be called from Python, using all default parameters:
+
+```
 import numpy as np
-import balsa
+import balsa_rfc
 
-# Create a dataset with two features per point
-dataset = np.array([[1, 1], [1, -1], [-1, -1], [-1, 1], [0, 0]], dtype=np.float64)
-labels = np.array([0, 1, 0, 1, 1], dtype=np.uint8)
+# Create a data set with two features per point.
+dataset = np.array(((1, 1), (1, -1), (-1, -1), (-1, 1), (0, 0)),dtype=np.float64)
 
-# Train a model and save it to 'model.balsa'
-balsa.train(dataset, labels, "model.balsa")
+# Create labels.
+labels = np.array((0, 1, 0, 1, 1), dtype=np.uint8)
 
-# Load the trained model
-classifier = balsa.RandomForestClassifier("model.balsa")
+# Train a model, write the output to 'model.balsa'.
+balsa_rfc.train(dataset, labels, "model.balsa")
 
-# Classify the dataset
+# Create a classifier from the trained random forest.
+classifier = balsa_rfc.RandomForestClassifier("model.balsa")
+
+# Classify the dataset.
 predictions = classifier.classify(dataset)
-print(f"Predictions: {predictions}")
 ```
 
-## Training Options
+### Additional Training Options
 
-### Basic Training Parameters
-```python
-balsa.train(
-    dataset,                      # Training data: shape [n_samples, n_features]
-    labels,                    # Ground truth labels: shape [n_samples]
-    "model.balsa",            # Path to save the trained model
-    features_to_consider=0,    # Features per split (0 = sqrt of total)
-    max_depth=4294967295,      # Maximum tree depth
-    min_purity=1.0,           # Minimum Gini-purity (1.0 = pure nodes)
-    tree_count=150,           # Number of trees in the forest
-    concurrent_trainers=1      # Number of parallel training threads
-)
+The trainer has several options that can be accessed through the constructor and
+`train()` method:
+
+```
+balsa_rfc.train(dataset, labels, "model.balsa",
+            features_to_consider = 0,
+            max_depth = 4294967295,
+            min_purity = 1.0,
+            tree_count = 150,
+            concurrent_trainers = 1)
 ```
 
-### Parameter Guide
+The purpose of these parameters is described in the docstring of `balsa.train`,
+as well as in the Balsa C++ documentation.
 
-**`features_to_consider`**: Number of features randomly selected when splitting a node
-- `0` (default): Uses sqrt(total_features)
[…]
