### 2025-11-13 7a17077
the performance tol needs the gnu time command to run
