### 2025-11-17 ef20f5d
adding a comment about custom installation paths in the main readme
