### 2025-11-17 3e7c775
adding a readme for the tool subfolder
