### 2025-11-17 51c9948
updating the tool readme wrong letter size
