### 2025-11-18 228566e
realistic example for the runtime analysis derived from TROPOMI
