### 2025-11-18 d8034fb
correcting the directory in the core readme file
