### 2025-11-19 ac61f6b
correcting an error in the python CMakeLists.txt
