### 2025-11-19 bf98e12
correcting inconsistent naming for the examples
