+++ core/Examples/classificationexample.cpp
@@ -6,7 +6,7 @@ using namespace balsa;
 int main( int, char ** )
 {
     // Read (and possibly convert) the data.
-    auto dataSet = readTableAs<double>( "fruit-data.balsa" );
+    auto dataSet = readTableAs<double>( "fruit-points.balsa" );
 
     // Classify the data.
     Table<Label>           labels( dataSet.getRowCount(), 1 );
