### 2025-11-19 e36ff29
update version of the joss paper
