+++ Paper/paper.bib
@@ -91,4 +91,13 @@ URL = {https://amt.copernicus.org/articles/16/1597/2023/},
 DOI = {10.5194/amt-16-1597-2023}
 }
 
-
+@Article{Wright:2017,
+  title = {{ranger}: A Fast Implementation of Random Forests for High Dimensional Data in {C++} and {R}},
+  author = {Marvin N. Wright and Andreas Ziegler},
+  journal = {Journal of Statistical Software},
+  year = {2017},
+  volume = {77},
+  number = {1},
+  pages = {1--17},
+  doi = {10.18637/jss.v077.i01},
+}
