+++ Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Balsa: A Fast C++ Random Forest Classifier with Commandline and Python Interface'
+title: 'Balsa: A Fast C++ Random Forest Classifier with Command-line and Python Interface'
 tags:
   - Python
   - C++
@@ -11,155 +11,192 @@ authors:
    - name: Tobias Borsdorff
      corresponding: true
      orcid: 0000-0002-4421-0187
-     equal-contribution: true
      affiliation: "1"
 
    - given-names: Denis 
      dropping-particle: de
      surname: Leeuw Duarte
-     equal-contribution: true
      affiliation: "2"
 
    - given-names: Joris
      dropping-particle: van
      surname: Zwieten
-     equal-contribution: true
      affiliation: "2"
 
    - name: Soumyajit Mandal
-     corresponding: true
      orcid: 0000-0002-2669-4798
-     equal-contribution: true
      affiliation: "1"
 
    - name: Jochen Landgraf
      orcid: 0000-0002-6069-0598
-     equal-contribution: true
      affiliation: "1"
 
-
-#  - name: Adrian M. Price-Whelan
-#    orcid: 0000-0000-0000-0000
-#    equal-contrib: true
-#    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-#  - name: Author Without ORCID
-#    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-#    affiliation: 2
-#  - name: Author with no affiliation
-#    corresponding: true # (This is how to denote the corresponding author)
-#    affiliation: 3
-#  - given-names: Ludwig
-#    dropping-particle: van
-#    surname: Beethoven
-#    affiliation: 3
 affiliations:
-  - name: SRON Netherlands Institute for Space Research, The Netherlands
+  - name: SRON Space Research Organization Netherlands
     index: 1
 
   - name: Jigsaw B.V., The Netherlands
     index: 2
 
-# - name: Lyman Spitzer, Jr. Fellow, Princeton University, United States
-#   index: 1
-#   ror: 00hx57361
-# - name: Institution Name, Country
-#   index: 2
-# - name: Independent Researcher, Country
-#   index: 3
-date: 10 Debember 2024
+date: 10 December 2024
 bibliography: paper.bib
-
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-#aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-#aas-journal: Astrophysical Journal <- The name of the AAS journal.
 
 ---
 
 # Summary
 
-A Random Forest classifier is a widely used machine learning method. It builds
-on the strengths of decision trees [@Pedregosa:2011] which are simple,
-intuitive models to classify input data into subsets based on the values of
-specific features. They form hierarchical structures with decision nodes
-leading to prediction [@Breiman:2001]. Combining the output of multiple
-decision trees improves its predictive accuracy, reduces the risk of
-overfitting, and effectively identifies outliers within data sets.  Balsa is a
-highly efficient C++ implementation of the Random Forest classifier concept,
-built with a focus on runtime and memory performance as key design priorities.
-Balsa provides multithreaded and distributed training capabilities, allowing
-users to scale machine learning processes across multiple computing cores or
-distributed systems by training separate Random Forests and combining them at
-the end. It supports both binary classification tasks and multi-label
-classification, expanding its use for a variety of machine learning challenges.
-Furthermore, Balsa includes comprehensive tools to assess the importance of
-features, prediction performance metrics, and statistical analysis, allowing
-users to gain insight into their data and model performance. To ensure ease of
[…]
