### 2025-11-25 a7ebe78
more realistic parameters for the python library in the readme
