### 2026-02-17 #8778 @HaoZeke
@vipinagrawal25 thanks for regenerating the proof — could you confirm whether the author's changes from November address the concerns you raised in your September review?

@bhaveshshrimali checking in again — you still have 6 unchecked items on the review checklist (reproducibility, functionality, performance, example usage, functionality docs, automated tests). Could you please provide an update on your progress? If you're no longer able to complete this review, please let me know so I can find a replacement.
