### 2026-03-04 SRON-Earth/Balsa#1 @bhaveshshrimali
Add community guidelines (CONTRIBUTING.md, CODE_OF_CONDUCT.md)
Hi @tobiasborsdorff @borsdorff, Please find me review/suggestions below

---

## Summary

The repository currently lacks community guidelines, which are required by the [JOSS review checklist](https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html). Specifically, there are no instructions for third parties on how to:

1. **Contribute** to the software
2. **Report issues** or problems with the software
3. **Seek support**

The only support-related information is a generic email address (`info@jigsaw.nl`) mentioned in the root README and the tools README. There is no `CONTRIBUTING.md`, `CODE_OF_CONDUCT.md`, or `.github/` directory with issue templates.

## What's needed

### 1. `CONTRIBUTING.md`

Please add a `CONTRIBUTING.md` file to the repository root that covers at a minimum:

- How to report bugs (e.g., open a GitHub issue)
- How to request features
- How to submit pull requests (coding style, testing expectations, branch strategy)
- How to seek support (email, GitHub Discussions, etc.)

Example structure:

```markdown
# Contributing to Balsa

Thank you for your interest in contributing to Balsa!

## Reporting Bugs

If you find a bug, please open an issue on the
[GitHub issue tracker](https://github.com/SRON-Earth/Balsa/issues) with:

- A clear description of the problem
- Steps to reproduce the issue
- Your operating system, compiler version, and Python version (if applicable)
- Any relevant error messages or logs

## Requesting Features

Feature requests are welcome! Please open a GitHub issue describing the feature
you'd like to see, why it would be useful, and any implementation ideas.

## Submitting Pull Requests

1. Fork the repository and create a new branch from `main`.
2. Make your changes, ensuring the code compiles and all tests pass
   (`balsa_test`).
3. Follow the existing code style (see `.clang-format` in the `core/` directory).
4. Submit a pull request with a clear description of your changes.

## Seeking Support

For general questions or support, you can:

- Open a [GitHub issue](https://github.com/SRON-Earth/Balsa/issues)
- Contact us at info@jigsaw.nl
```

### 2. `CODE_OF_CONDUCT.md`

Please add a `CODE_OF_CONDUCT.md` to the repository root. A common choice is the [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/), which is widely used in open-source projects:

```markdown
# Contributor Covenant Code of Conduct

## Our Pledge

We as members, contributors, and leaders pledge to make participation in our
community a harassment-free experience for everyone...

(full text at https://www.contributor-covenant.org/version/2/1/code_of_conduct/)
```

## Context

This is a review requirement for the JOSS submission:
[openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

The review checklist item reads:

> **Community guidelines:** Are there clear guidelines for third parties wishing to
> 1. Contribute to the software
> 2. Report issues or problems with the software
> 3. Seek support

This is currently the **only unchecked item** in my review checklist.

---
