### 2026-03-04 SRON-Earth/Balsa#2 @bhaveshshrimali
Fix typo in pyproject.toml URLs: `SRON-Eath` → `SRON-Earth`
## Summary

All four project URLs in [`python/pyproject.toml`](https://github.com/SRON-Earth/Balsa/blob/main/python/pyproject.toml) contain a typo: `SRON-Eath` instead of `SRON-Earth`.

## Current (broken)

```toml
[project.urls]
Homepage = "https://github.com/SRON-Eath/Balsa"
Documentation = "https://github.com/SRON-Eath/Balsa/blob/main/README.md"
Repository = "https://github.com/SRON-Eath/Balsa"
Issues = "https://github.com/SRON-Eath/Balsa/issues"
```

These URLs currently lead to a 404 page on GitHub.

## Expected (fixed)

```toml
[project.urls]
Homepage = "https://github.com/SRON-Earth/Balsa"
Documentation = "https://github.com/SRON-Earth/Balsa/blob/main/README.md"
Repository = "https://github.com/SRON-Earth/Balsa"
Issues = "https://github.com/SRON-Earth/Balsa/issues"
```

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
