### 2026-03-04 SRON-Earth/Balsa#3 @bhaveshshrimali
Fix `paper.bib`: unused `balsagit` entry points to wrong repository URL
## Summary

The file [`Paper/paper.bib`](https://github.com/SRON-Earth/Balsa/blob/main/Paper/paper.bib) contains an entry `balsagit` that has two problems:

### Problem 1: URL points to a different repository owner

The entry currently points to `https://github.com/borsdorff/Balsa`, but the canonical repository is `https://github.com/SRON-Earth/Balsa`.

```bibtex
@misc{balsagit,
  author = {T. Borsdorff and D. de Leeuw {Duarte} and J.E. van Zwieten and J. Landgraf},
  title = {Balsa: A Fast C++ Random Forest Classifier},
  year = {2024},
  publisher = {GitHub},
  journal = {GitHub repository},
  url = {https://github.com/borsdorff/Balsa}   % <-- wrong owner
}
```

### Problem 2: Entry is never cited in `paper.md`

A search of `paper.md` finds no `[@balsagit]` citation. The entry appears to be vestigial.

### Suggested fix

Either:

**Option A** — Remove the entry entirely if it's not needed (the repository URL is already linked in the "Availability" section of the paper).

**Option B** — Update the URL and cite it in the paper:

```bibtex
@misc{balsagit,
  author = {T. Borsdorff and D. de Leeuw {Duarte} and J.E. van Zwieten and J. Landgraf},
  title = {Balsa: A Fast C++ Random Forest Classifier},
  year = {2024},
  publisher = {GitHub},
  journal = {GitHub repository},
  url = {https://github.com/SRON-Earth/Balsa}
}
```

And then add `[@balsagit]` somewhere in `paper.md`, e.g., in the Availability section.

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
