### 2026-03-04 SRON-Earth/Balsa#4 @bhaveshshrimali
`rfcperf` fails on macOS: `time -v` is not supported by BSD `time` [minor]
## Summary

The `rfcperf` benchmarking tool uses GNU `time -v` to collect system resource usage statistics. On macOS, the built-in `/usr/bin/time` is BSD `time`, which does not support the `-v` flag, causing all benchmark runs to fail with:

```
time: illegal option -- v
usage: time [-al] [-h | -p] [-o file] utility [argument ...]
```

## Steps to reproduce

```bash
# On macOS
cd tools
python3 -m rfcperf profile my_dataset.json balsa -n 1000 -e 50 -t 4 -p 33 -d 20
```

Output:

```
Running balsa using 4 threads...
1000
Run failed: 'Program '.../balsa_train' failed with exit code: 1'.
```

The error in the run log is:

```
time: illegal option -- v
```

## Root cause

In [`tools/rfcperf/util.py` line 9](https://github.com/SRON-Earth/Balsa/blob/main/tools/rfcperf/util.py#L9):

```python
result = subprocess.run(["time", "-v", "-o", time_file, str(program), *args],
                        capture_output=True, text=True,
                        timeout=timeout, cwd=cwd)
```

This hardcodes `"time"`, which resolves to BSD `time` on macOS.

## Workaround

Install GNU time via Homebrew and add it to PATH:

```bash
brew install gnu-time
export PATH="/opt/homebrew/opt/gnu-time/libexec/gnubin:$PATH"
```

## Suggested fix

Update `util.py` to auto-detect the platform and use `gtime` on macOS:

```python
import platform
import shutil

def _get_time_command():
    """Return the GNU time command for the current platform."""
    if platform.system() == "Darwin":
        # macOS: GNU time is typically installed as 'gtime' via Homebrew
        if shutil.which("gtime"):
            return "gtime"
        raise RuntimeError(
            "GNU time is required but not found. "
            "Install it with: brew install gnu-time"
        )
    return "time"

def run_program(program, *args, log=False, log_prefix=None, time_file=None,
                timeout=None, cwd=None):
    if time_file is not None:
        time_cmd = _get_time_command()
        result = subprocess.run(
            [time_cmd, "-v", "-o", time_file, str(program), *args],
            capture_output=True, text=True, timeout=timeout, cwd=cwd)
    else:
        # ... existing code ...
```

Alternatively, document the macOS workaround in `tools/rfcperf/README.md` under the Requirements section, e.g.:

```markdown
**Requirements:**
- Python ≥ 3.7
- NumPy
- matplotlib
- jsonpickle
- GNU time command
  - **Linux:** typically pre-installed (part of coreutils)
  - **macOS:** install via `brew install gnu-time` and add to PATH:
    `export PATH="/opt/homebrew/opt/gnu-time/libexec/gnubin:$PATH"`
```

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
