### 2026-03-04 SRON-Earth/Balsa#5 @bhaveshshrimali
[BUG] `balsa_classify` crashes (segfault) when model and data have different feature counts
## Summary

When `balsa_classify` is given a model trained on *N* features and input data with a different number of features, the tool crashes with a **segmentation fault** (exit code 139) instead of producing a helpful error message. The Python bindings (`balsa_rfc`) correctly raise a `ValueError` in this situation, but the C++ CLI tool lacks the equivalent check.

## Steps to reproduce

```bash
# Train a 3-feature model
balsa_generate -p 1000 some_3feat.conf data3.balsa labels3.balsa
balsa_train -c 50 data3.balsa labels3.balsa model_3feat.balsa

# Classify data with 7 features using the 3-feature model → SEGFAULT
balsa_classify model_3feat.balsa data_7feat.balsa
# Output:
#   Ingesting data...
#   Dataset loaded: 7 features x N points.
#   zsh: segmentation fault  balsa_classify model_3feat.balsa data_7feat.balsa
#   Exit code: 139
```

This also crashes when the data feature count is a multiple of the model feature count (e.g., 6-feature data with 2-feature model), because the modulo check in the internal classify method doesn't catch it.

## Root cause

In the CLI tool source code (`core/Sources/classify.cpp`), the data is loaded and passed to the classifier without comparing `dataSet.getColumnCount()` against `classifier.getFeatureCount()`. The internal classify methods only perform a modulo check on the flat iterator range, which is insufficient when data is loaded as a 2D table.

## Python bindings handle this correctly

For comparison, the Python binding in `balsamodule.cpp` explicitly checks:

```cpp
if ( number_of_features != self->m_classifier->getFeatureCount() )
{
    PyErr_SetString( PyExc_ValueError,
        "The number of features in the input data differs from the number "
        "of features expected by the classifier." );
    ...
}
```

## Suggested fix

Add a check in `classify.cpp` between loading the data and calling classify:

```cpp
if ( dataSet.getColumnCount() != classifier.getFeatureCount() )
    throw ClientError( "Feature count mismatch: data has "
        + std::to_string( dataSet.getColumnCount() ) + " features, model expects "
        + std::to_string( classifier.getFeatureCount() ) + "." );
```

## Impact

This could affect any user who accidentally passes mismatched data, or pipelines where file paths get mixed up. A crash with no error message is difficult to debug.

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
