### 2026-03-04 SRON-Earth/Balsa#6 @bhaveshshrimali
rfcperf README: data size description and thread count mismatch in example
## Summary

In [`tools/rfcperf/README.md`](https://github.com/SRON-Earth/Balsa/blob/main/tools/rfcperf/README.md), the "Run Profiling" example contains two inconsistencies between the command and its description.

### Problem 1: Data sizes are described incorrectly (off by 10×)

The command specifies:

```
-n 1000,250000,500000,750000,1000000
```

These are **1K, 250K, 500K, 750K, and 1M** samples.

But the accompanying text says:

> This profiles Balsa, scikit-learn and the ranger library on datasets of **1K, 2.5M, 5M, 7.5M, and 10M** samples

The description is **off by a factor of 10** (250K ≠ 2.5M, etc.).

### Problem 2: Thread count mismatch

The command specifies `-t 8` (8 threads) but the description says **"16 threads"**.

### Suggested fix

Either update the description to match the command:

> This profiles Balsa, scikit-learn and the ranger library on datasets of 1K, 250K, 500K, 750K, and 1M samples, using 150 trees, **8** threads, and a 33% test split.

Or update the command to match the description, if the larger sizes were intended.

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
