### 2026-03-04 SRON-Earth/Balsa#7 @bhaveshshrimali
[Enhancement]: rfcperf classification metrics only work for binary classification (labels 0 and 1)
## Summary

The `get_classification_scores()` function in [`tools/rfcperf/util.py`](https://github.com/SRON-Earth/Balsa/blob/main/tools/rfcperf/util.py) is hardcoded to compute confusion matrix entries for binary classification with labels `0.0` and `1.0` only:

```python
num_true_positives  = np.sum((predicted_labels == 1.0) & (labels == 1.0))
num_false_positives = np.sum((predicted_labels == 1.0) & (labels == 0.0))
num_true_negatives  = np.sum((predicted_labels == 0.0) & (labels == 0.0))
num_false_negatives = np.sum((predicted_labels == 0.0) & (labels == 1.0))
```

Similarly, `generate_datasets()` in `data.py` asserts:

```python
assert np.sum(labels == 0.0) + np.sum(labels == 1.0) == labels.size
```

This means rfcperf **cannot be used to benchmark multi-class classification**, even though the Balsa library itself fully supports multi-class classification.

## Impact

- Users who try to profile multi-class datasets will get incorrect metrics (only labels matching 0 or 1 are counted) without any warning or error.
- The assertion in `data.py` will halt execution with an `AssertionError` for multi-class datasets, but without a helpful error message.

## Suggested fix

At minimum, add a clear error or warning message when multi-class data is detected:

```python
unique_labels = np.unique(labels)
if not np.array_equal(unique_labels, [0.0, 1.0]):
    raise ValueError(f"rfcperf currently only supports binary classification "
                     f"(labels 0 and 1). Found labels: {unique_labels}")
```

Ideally, extend the scoring to support multi-class classification (e.g., using sklearn-style macro/micro averaging).

## Context

This is not a blocking issue for the JOSS review since the paper's TROPOMI dataset is binary, but it limits the tool's generality as described in the documentation.

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)

---

---
