### 2026-03-04 SRON-Earth/Balsa#8 @bhaveshshrimali
[Documentation]: Full figure reproduction requires Ranger, which is not bundled or automated
## Summary

The paper's Figures 1–3 compare runtime, memory, and accuracy of **three** classifiers: Balsa, scikit-learn, and **Ranger**. The `rfcperf` tool supports Ranger via its driver, but:

1. Ranger must be **separately downloaded and compiled** from https://github.com/imbs-hl/ranger
2. The rfcperf README mentions this only briefly: *"You need to install sklearn in your python environment and download and built ranger a cpp implementation of the random forest classifier concept."*
3. There are no build instructions for Ranger, no version pinned, and no automated way to set it up.

This means a reviewer (or reader) attempting to reproduce the paper's three-classifier comparison cannot do so with just the Balsa repository alone. While Balsa vs. sklearn is readily reproducible (and I confirmed this during testing), including Ranger requires external setup.

## Suggestion

Consider one of:

1. **Document the Ranger setup** more explicitly in the rfcperf README (which version to use, build instructions for macOS/Linux, how to configure the path).
2. **Provide a script** that downloads and builds Ranger automatically.
3. At minimum, make it clear in the paper or README which figures can be reproduced with just Balsa + sklearn, and that Ranger is an optional third comparison.

## Reproduction result

I was able to reproduce Figures 1–3 for Balsa vs. sklearn (2 of 3 classifiers) using the `rfcperf` tool. The results are consistent with the paper's claims:
- Balsa trains and classifies significantly faster than sklearn
- Balsa uses significantly less memory
- Both achieve comparable accuracy

## Context

Found during JOSS review: [openjournals/joss-reviews#8778](https://github.com/openjournals/joss-reviews/issues/8778)
