### 2026-03-17 4b7f374
Updated the README to reflect the working directory for rfcperf commands (fixes #10)
