### 2026-03-17 be7c06e
Updated the README to reflect the ldconfig step after installation (fixes #9)
