### 2026-03-17 1c3447b
Added the CONTRIBUTING.md and CODE_OF_CONDUCT.md files following the guidelines (fixes #1)
