+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,8 @@
+# Contributor Covenant Code of Conduct
+
+## Our Pledge
+
+We as members, contributors, and leaders pledge to make participation in our
+community a harassment-free experience for everyone...
+
+(full text at https://www.contributor-covenant.org/version/2/1/code_of_conduct/)
