+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+# Contributing to Balsa
+
+Thank you for your interest in contributing to Balsa!
+
+## Reporting Bugs
+
+If you find a bug, please open an issue on the
+[GitHub issue tracker](https://github.com/SRON-Earth/Balsa/issues) with:
+
+- A clear description of the problem
+- Steps to reproduce the issue
+- Your operating system, compiler version, and Python version (if applicable)
+- Any relevant error messages or logs
+
+## Requesting Features
+
+Feature requests are welcome! Please open a GitHub issue describing the feature
+you'd like to see, why it would be useful, and any implementation ideas.
+
+## Submitting Pull Requests
+
+1. Fork the repository and create a new branch from `main`.
+2. Make your changes, ensuring the code compiles and all tests pass
+   (`balsa_test`).
+3. Follow the existing code style (see `.clang-format` in the `core/` directory).
+4. Submit a pull request with a clear description of your changes.
+
+## Seeking Support
+
+For general questions or support, you can:
+
+- Open a [GitHub issue](https://github.com/SRON-Earth/Balsa/issues)
+- Contact us at info@jigsaw.nl
