### 2026-03-17 6a65042
Typo fixed in the pyproject.toml file (fixes #2)
