### 2026-03-17 d7cd7cc
unused balsagit citation is removed from paper.bib (fixes #3)
