+++ Paper/paper.bib
@@ -21,15 +21,6 @@ doi          = {10.5281/zenodo.14186320},
 url          = {https://doi.org/10.5281/zenodo.14186320}
 }
 
-@misc{balsagit,
-  author = {T. Borsdorff and D. de Leeuw {Duarte} and J.E. van Zwieten and J. Landgraf},
-  title = {Balsa: A Fast C++ Random Forest Classifier},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/borsdorff/Balsa}
-}
-
 
 @manual{Borsdorff:2024_pum,
 author       = {Borsdorff, Tobias and Mandal, Soumyajit and Martinez Velarte, Mari and Barr, Andrew and Landgraf, Jochen},
