### 2026-03-17 550fdb9
Updated util.py to auto-detect platform and use gtime on macOS (fixes #4)
