### 2026-03-17 c6fb4b6
Description in rfcperf README.md updated to match the example  (fixes #6)
