### 2026-03-17 3e5fb02
Error message is added when multi-class data is detected. We will update the script in future to support multi-class data.  (fixes #7)
