### 2026-03-17 700da26
To prevent the segfault in balsa_classify we add the suggested check (fixes #5)
