### 2026-03-17 ee1db7f
We documented the Ranger setup,config,version in the README (fixes #8)
