### 2026-03-17 #8778 @tobiasborsdorff
Dear @bhaveshshrimali, @vipinagrawal25, and @HaoZeke,

Thank you all for your thorough and constructive review of our software package. Your detailed feedback has led to meaningful improvements to the code and documentation, and I am grateful for the time and effort you invested in the process.

I am pleased to inform you that all open issues have been addressed and closed.

Best regards, Tobias
