### 2026-08-18 db73ed5
some corrections to paper.bib and paper.md
