+++ Paper/paper.bib
@@ -1,6 +1,6 @@
 
 @Article{Borsdorff:2024_paper,
-AUTHOR = {Borsdorff, Tobias and Martinez-Velarte, Mari C. and Sneep, Maarten and ter Linden, Mark and Landgraf, Jochen},
+AUTHOR = {Borsdorff, Tobias and Martinez-Velarte, Mari C. and Sneep, Maarten and {ter Linden}, Mark and Landgraf, Jochen},
 TITLE = {Random Forest Classifier for Cloud Clearing of the Operational TROPOMI XCH4 Product},
 JOURNAL = {Remote Sensing},
 VOLUME = {16},
@@ -13,7 +13,7 @@ DOI = {10.3390/rs16071208}
 }
 
 @manual{Borsdorff:2024_atbd,
-title        = {Algorithm Theoretical Baseline Document for Sentinel-5 Precursor: TROPOMI CH4 Random Forest Cloud Filter},
+title        = {Algorithm Theoretical Baseline Document for {Sentinel-5 Precursor}: TROPOMI CH4 Random Forest Cloud Filter},
 author       = {Borsdorff, Tobias and Mandal, Soumyajit and Martinez Velarte, Mari and Barr, Andrew and Landgraf, Jochen},
 month        = dec,
 year         = 2024,
@@ -24,7 +24,7 @@ url          = {https://doi.org/10.5281/zenodo.14186320}
 
 @manual{Borsdorff:2024_pum,
 author       = {Borsdorff, Tobias and Mandal, Soumyajit and Martinez Velarte, Mari and Barr, Andrew and Landgraf, Jochen},
-title        = {Product User Manual for the Random Forest Classifier (RFC) C++ Implementation to be used for TROPOMI CH4},
+title        = {Product User Manual for the Random Forest Classifier (RFC) {C++} Implementation to be used for TROPOMI CH4},
 month        = dec,
 year         = 2024,
 publisher    = {Zenodo},
@@ -59,7 +59,7 @@ doi          = {10.1023/A:1010933404324}
 }
 
 @Article{Lorente:2021,
-AUTHOR = {Lorente, A. and Borsdorff, T. and Butz, A. and Hasekamp, O. and aan de Brugh, J. and Schneider, A. and Wu, L. and Hase, F. and Kivi, R. and Wunch, D. and Pollard, D. F. and Shiomi, K. and Deutscher, N. M. and Velazco, V. A. and Roehl, C. M. and Wennberg, P. O. and Warneke, T. and Landgraf, J.},
+AUTHOR = {Lorente, A. and Borsdorff, T. and Butz, A. and Hasekamp, O. and {aan de Brugh}, J. and Schneider, A. and Wu, L. and Hase, F. and Kivi, R. and Wunch, D. and Pollard, D. F. and Shiomi, K. and Deutscher, N. M. and Velazco, V. A. and Roehl, C. M. and Wennberg, P. O. and Warneke, T. and Landgraf, J.},
 TITLE = {Methane retrieved from TROPOMI: improvement of the data product and validation of the first 2 years of measurements},
 JOURNAL = {Atmospheric Measurement Techniques},
 VOLUME = {14},
