### 2026-08-18 #8778 @tobiasborsdorff
@crvernon - thanks for the detailed feedback. I've addressed all the items:

**Archive:**
- Updated the Zenodo archive metadata title to match the paper exactly, removing the "SRON-Earth/Balsa: " prefix.

**Paper:**
- Lines 52-55: Fixed the duplicated citation - now reads "as described by Borsdorff, Martinez-Velarte, et al. (2024)."
- Line 96: Changed "Authors Contribution" to "Author contributions."
- Lines 106-108: "Sentinel-5 Precursor" is now protected with curly brackets in the bib file and renders correctly capitalized.
- Lines 109-112: "C++" is now protected with curly brackets in the bib file and renders correctly capitalized.
- Lines 113-115: Fixed to "ter Linden, M." by wrapping the full family name in curly brackets in the bib entry.
- Lines 118-123: Fixed to "aan de Brugh, J." the same way.

All changes are pushed to the repository. Let me know if anything else is needed.
