### 2026-08-18 #8778 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
