### 2026-08-19 #8778 @vipinagrawal25
Thanks Everyone for the opportunity to review the paper. I enjoyed the process.

Regards
Vipin

> On Aug 19, 2026, at 8:25 AM, Tobias Borsdorff ***@***.***> wrote:
>
>
> tobiasborsdorff
>  left a comment
> (openjournals/joss-reviews#8778)
>  <https://github.com/openjournals/joss-reviews/issues/8778#issuecomment-5342777202>
> Thank you so much, @crvernon <https://github.com/crvernon>! It was a great experience, and I really appreciate all the time, effort, and expertise that @HaoZeke <https://github.com/HaoZeke>, @vipinagrawal25 <https://github.com/vipinagrawal25>, and @bhaveshshrimali <https://github.com/bhaveshshrimali> put into the review and publication process.
>
